refactor benchmarking script
