+++ isaaclab_arena/relations/benchmark/__init__.py
@@ -1,0 +1,62 @@
+# Copyright (c) 2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Curated public API for relation solver benchmarks."""
+
+from isaaclab_arena.relations.benchmark.environment import run_environment_benchmark
+from isaaclab_arena.relations.benchmark.execution import run_benchmarks
+from isaaclab_arena.relations.benchmark.models import (
+    BenchmarkMeasurement,
+    BenchmarkRun,
+    BenchmarkScenario,
+    BenchmarkStatus,
+    BenchmarkTarget,
+    Clock,
+    CollisionModeName,
+    DeviceMetadata,
+)
+from isaaclab_arena.relations.benchmark.reporting import (
+    build_distributed_run,
+    build_run,
+    format_results_table,
+    requested_scenario_ids,
+    search_capacity,
+    write_results_csv,
+    write_results_json,
+)
+from isaaclab_arena.relations.benchmark.solver import (
+    default_scenarios,
+    env_count_sweep,
+    object_count_sweep,
+    run_placer_benchmark,
+    run_solver_benchmark,
+    scenarios_for_modes,
+)
+
+__all__ = [
+    "BenchmarkMeasurement",
+    "BenchmarkRun",
+    "BenchmarkScenario",
+    "BenchmarkStatus",
+    "BenchmarkTarget",
+    "Clock",
+    "CollisionModeName",
+    "DeviceMetadata",
+    "build_distributed_run",
+    "build_run",
+    "default_scenarios",
+    "env_count_sweep",
+    "format_results_table",
+    "object_count_sweep",
+    "requested_scenario_ids",
+    "run_benchmarks",
+    "run_environment_benchmark",
+    "run_placer_benchmark",
+    "run_solver_benchmark",
+    "scenarios_for_modes",
+    "search_capacity",
+    "write_results_csv",
+    "write_results_json",
+]
